+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 Added support for `no_border` to `windowstate`
 Added support for `minimized` to `windowstate`
 
+New window actions:
+
+- `windowmaximize`
+
 ## v0.2.1 (2023-11-23)
 
 Reduced binary size.